// frontend/src/lib/types.ts
export enum Role {
  LAWYER = "LAWYER",
  CLIENT = "CLIENT",
}
// ...
export interface AppointmentsType {
  id?: string;
  client_id?: string;
  lawyer_id?: string;
  description: string;
  date: Date;
  status?: APPOINTMENT_STATUS;
  createdAt?: Date;
  updatedAt?: Date;
  role?: Role;
  client?: {
    firstName: string;
    lastName: string;
    email: string;
    phone_number: string;
  };
  lawyer?: {
    user: {
      firstName: string;
      lastName: string;
      email: string;
      phone_number: string;
    };
  };
}

export enum APPOINTMENT_STATUS {
  PENDING = "PENDING",
  CONFIRMED = "CONFIRMED",
  COMPLETED = "COMPLETED",
  CANCELED = "CANCELED",
}